+++ schema/sdss5db/opsdb/opsdb.sql
@@ -138,6 +138,29 @@ CREATE TABLE opsdb.overhead(
     elapsed REAL,
     success BOOLEAN);
 
+CREATE TABLE opsdb.pred_snr (
+    pk SERIAL PRIMARY KEY NOT NULL,
+    robodamus_version TEXT,
+    model_id TEXT,
+    pred_time TIMESTAMP,
+    camera_pk SMALLINT,
+    design_mode_label TEXT,
+    pred_type TEXT,
+    pred_value REAL,
+    pred_flag INTEGER,
+    num_gfas INTEGER,
+    gfa_expid INTEGER,
+    gfa_date_obs TIMESTAMP);
+
+CREATE TABLE opsdb.pred_snr_par (
+    pk SERIAL PRIMARY KEY NOT NULL,
+    pred_snr_pk BIGINT,
+    label TEXT,
+    datatype TEXT,
+    value_f DOUBLE PRECISION,
+    value_i BIGINT,
+    value_s TEXT);
+
 -- Foreign keys
 
 ALTER TABLE ONLY opsdb.base_priority
@@ -239,6 +262,19 @@ ALTER TABLE ONLY opsdb.overhead
     ON UPDATE CASCADE ON DELETE CASCADE
     DEFERRABLE INITIALLY DEFERRED;
 
+ALTER TABLE ONLY opsdb.pred_snr
+    ADD CONSTRAINT camera_fk
+    FOREIGN KEY (camera_pk) REFERENCES opsdb.camera(pk);
+
+ALTER TABLE ONLY opsdb.pred_snr
+    ADD CONSTRAINT design_mode_fk
+    FOREIGN KEY (design_mode_label) REFERENCES targetdb.design_mode(label);
+
+ALTER TABLE ONLY opsdb.pred_snr_par
+    ADD CONSTRAINT pred_snr_fk
+    FOREIGN KEY (pred_snr_pk) REFERENCES opsdb.pred_snr(pk);
+
+
 -- Table data
 
 INSERT INTO opsdb.exposure_flavor VALUES
@@ -294,6 +330,42 @@ CREATE INDEX CONCURRENTLY qr_exposure_pk_idx
 CREATE INDEX CONCURRENTLY overhead_configuration_id_idx
     ON opsdb.overhead
     USING BTREE(configuration_id);
+
+CREATE INDEX CONCURRENTLY pred_snr_camera_pk_idx
+    ON opsdb.pred_snr
+    USING BTREE(camera_pk);
+
+CREATE INDEX CONCURRENTLY pred_snr_design_mode_label_idx
+    ON opsdb.pred_snr
+    USING BTREE(design_mode_label);
+
+CREATE INDEX CONCURRENTLY pred_snr_pred_time_idx
+    ON opsdb.pred_snr
+    USING BTREE(pred_time);
+
+CREATE INDEX CONCURRENTLY pred_snr_pred_type_idx
+    ON opsdb.pred_snr
+    USING BTREE(pred_type);
+
+CREATE INDEX CONCURRENTLY pred_snr_gfa_date_obs_idx
+    ON opsdb.pred_snr
+    USING BTREE(gfa_date_obs);
+
+CREATE INDEX CONCURRENTLY pred_snr_model_id_idx
+    ON opsdb.pred_snr
+    USING BTREE(model_id);
+
+CREATE INDEX CONCURRENTLY pred_snr_robodamus_version_idx
+    ON opsdb.pred_snr
+    USING BTREE(robodamus_version);
+
+CREATE INDEX CONCURRENTLY pred_snr_par_pred_snr_pk_idx
+    ON opsdb.pred_snr_par
+    USING BTREE(pred_snr_pk);
+
+CREATE INDEX CONCURRENTLY pred_snr_label_idx
+    ON opsdb.pred_snr_par
+    USING BTREE(label);
 
 -- pop function to retrieve next in queue and increment